trying to render switches
